Define real precision explicitly to prevent mismatch errors

+++ src/cosp_control/cosp_reff_mod.F90
@@ -43,16 +43,16 @@ LOGICAL,INTENT(IN) :: flux
 ! X1,X3: n_ax = X1*exp(-X3*T[degC])
 ! X2: n_bx
 ! X4: alpha_x
-REAL,INTENT(IN) :: x1
-REAL,INTENT(IN) :: x2
-REAL,INTENT(IN) :: x3
-REAL,INTENT(IN) :: x4
+REAL(wp),INTENT(IN) :: x1
+REAL(wp),INTENT(IN) :: x2
+REAL(wp),INTENT(IN) :: x3
+REAL(wp),INTENT(IN) :: x4
 ! These are the constants that define the mass-diameter relationship.
 ! M_x(D) = a_x*D^(b_x)
 ! A: a_x
 ! B: b_x
-REAL,INTENT(IN) :: a
-REAL,INTENT(IN) :: b
+REAL(wp),INTENT(IN) :: a
+REAL(wp),INTENT(IN) :: b
 ! These are the constants that define the terminal fall speed relationship.
 ! V_x(D) = c_x*D^(d_x)*exp(-h_x)*(rho_0/rho)^g_x
 ! C: c_x
@@ -60,29 +60,29 @@ REAL,INTENT(IN) :: b
 ! G: g_x
 ! h_x is always 0
 ! Abel and Shipway (2007) is not supported
-REAL,INTENT(IN) :: c
-REAL,INTENT(IN) :: d
-REAL,INTENT(IN) :: g
+REAL(wp),INTENT(IN) :: c
+REAL(wp),INTENT(IN) :: d
+REAL(wp),INTENT(IN) :: g
 ! Dimensions
 INTEGER,INTENT(IN) :: npoints
 INTEGER,INTENT(IN) :: model_levels
 ! Air temperature [K]
-REAL,INTENT(IN) :: t(npoints,model_levels)
+REAL(wp),INTENT(IN) :: t(npoints,model_levels)
 ! Air density [kg/m^3]
-REAL,INTENT(IN) :: rho(npoints,model_levels)
+REAL(wp),INTENT(IN) :: rho(npoints,model_levels)
 ! Hydrometeor mixing ratio [kg/kg]
-REAL,INTENT(IN) :: mr(npoints,model_levels)
+REAL(wp),INTENT(IN) :: mr(npoints,model_levels)
 
 !----Output arguments
 ! Effective radius [m]
-REAL,INTENT(OUT) :: Reff(npoints,model_levels)
+REAL(wp),INTENT(OUT) :: Reff(npoints,model_levels)
 
 !----Local variables
-REAL,PARAMETER :: rho_0 = 1.0
-REAL,PARAMETER :: zerodegc = 273.15
-REAL,PARAMETER :: t_agg_min = -45.0
-REAL :: gamma_a3,gamma_a4,gamma_ab1,gamma_abd1,frac_exp,gamma_ratio
-REAL :: F_nax(npoints,model_levels)
+REAL(wp),PARAMETER :: rho_0 = 1.0_wp
+REAL(wp),PARAMETER :: zerodegc = 273.15_wp
+REAL(wp),PARAMETER :: t_agg_min = -45.0_wp
+REAL(wp) :: gamma_a3,gamma_a4,gamma_ab1,gamma_abd1,frac_exp,gamma_ratio
+REAL(wp) :: F_nax(npoints,model_levels)
 CHARACTER(LEN=*), PARAMETER :: RoutineName='COSP_REFF'
 INTEGER :: icode,i,k
 INTEGER(KIND=jpim), PARAMETER :: zhook_in  = 0
@@ -92,31 +92,31 @@ REAL(KIND=jprb)               :: zhook_handle
 IF (lhook) CALL dr_hook(ModuleName//':'//RoutineName,zhook_in,zhook_handle)
 icode = 9
 
-CALL gammafunc(3.0+x4,gamma_a3)
-CALL gammafunc(4.0+x4,gamma_a4)
-CALL gammafunc(1.0+x4+b,gamma_ab1)
-CALL gammafunc(1.0+x4+b+d,gamma_abd1)
+CALL gammafunc(3.0_wp+x4,gamma_a3)
+CALL gammafunc(4.0_wp+x4,gamma_a4)
+CALL gammafunc(1.0_wp+x4+b,gamma_ab1)
+CALL gammafunc(1.0_wp+x4+b+d,gamma_abd1)
 
 
-Reff = 0.0
+Reff = 0.0_wp
 
-IF (a <= 0.0) CALL Ereport(RoutineName, icode," A <= 0.0")
+IF (a <= 0.0_wp) CALL Ereport(RoutineName, icode," A <= 0.0")
 
 ! Compute intercept as function of T, if needed
-IF (x3 /= 0.0) F_nax = EXP(-x3*MAX(t-zerodegc,t_agg_min))
+IF (x3 /= 0.0_wp) F_nax = EXP(-x3*MAX(t-zerodegc,t_agg_min))
 
 ! Compute the parameter lambda^-1 of the PSD. stored in variable Reff
 IF (flux) THEN ! precipitation flux. Fall speed needed
-  frac_exp = 1.0/(x4+d-x2+4.0)
-  IF (c <= 0.0) CALL Ereport(RoutineName, icode," C <= 0.0")
-  IF (x3 /= 0.0) THEN
+  frac_exp = 1.0_wp/(x4+d-x2+4.0_wp)
+  IF (c <= 0.0_wp) CALL Ereport(RoutineName, icode," C <= 0.0")
+  IF (x3 /= 0.0_wp) THEN
     Reff = (mr/(a*c*((rho_0/rho)**g)*x1*F_nax*gamma_abd1))**frac_exp
   ELSE
     Reff = (mr/(a*c*((rho_0/rho)**g)*x1*gamma_abd1))**frac_exp
   END IF
 ELSE ! mixing ratio
-  frac_exp = 1.0/(x4+b-x2+1.0)
-  IF (x3 /= 0.0) THEN
+  frac_exp = 1.0_wp/(x4+b-x2+1.0_wp)
+  IF (x3 /= 0.0_wp) THEN
     Reff = ((rho*mr)/(a*gamma_ab1*x1*F_nax))**frac_exp
   ELSE
     Reff = ((rho*mr)/(a*gamma_ab1*x1))**frac_exp
@@ -124,14 +124,14 @@ ELSE ! mixing ratio
 END IF
 
 ! Compute radius and apply sanity check
-gamma_ratio = 0.5*(gamma_a4/gamma_a3)
+gamma_ratio = 0.5_wp*(gamma_a4/gamma_a3)
 !$OMP PARALLEL DEFAULT(NONE) PRIVATE(i,k)                                      &
 !$OMP SHARED(model_levels, npoints, Reff, gamma_ratio)
 !$OMP DO SCHEDULE(STATIC)
 DO k = 1, model_levels
   DO i = 1, npoints
-    IF (Reff(i,k) > 0.0) Reff(i,k) = gamma_ratio*Reff(i,k)
-    IF (Reff(i,k) < 0.0) Reff(i,k) = 0.0
+    IF (Reff(i,k) > 0.0_wp) Reff(i,k) = gamma_ratio*Reff(i,k)
+    IF (Reff(i,k) < 0.0_wp) Reff(i,k) = 0.0_wp
   END DO
 END DO
 !$OMP END DO
@@ -144,10 +144,10 @@ subroutine gammafunc(y_in,gam_out)
 use yomhook, only: lhook, dr_hook
 use parkind1, only: jprb, jpim
 implicit none
-real ::                                                      &
+real(wp) ::                                                      &
                             !, intent(in)
   y_in
-real ::                                                      &
+real(wp) ::                                                      &
                             !, intent(out)
   gam_out
 ! Gamma function of Y
@@ -169,7 +169,7 @@ character(len=*), parameter :: RoutineName='GAMMAF'
 if (lhook) call dr_hook(ModuleName//':'//RoutineName,zhook_in,zhook_handle)
 
 y=y_in
-gg=1.0
+gg=1.0_wp
 m=floor(y)
 x=y-m
 if (m > 1) then
@@ -183,8 +183,8 @@ else if (m < 1) then
     gg=gg/g
   end do
 end if
-pare=-0.5748646*x+0.9512363*x*x-0.6998588*x*x*x                                &
-+0.4245549*x*x*x*x-0.1010678*x*x*x*x*x+1.0
+pare=-0.5748646_wp*x+0.9512363_wp*x*x-0.6998588_wp*x*x*x &
++0.4245549_wp*x*x*x*x-0.1010678_wp*x*x*x*x*x+1.0_wp
 gam=pare*gg
 gam_out=gam
 if (lhook) call dr_hook(ModuleName//':'//RoutineName,zhook_out,zhook_handle)
